Created separated component as link item and added click link to it
